Update README to clearly state this repo is only for mods of the example
Only the example is intended to be used from this repository and not the main development files.  Please be sure to run `npm install` in the examples/express2 directory to get the most recent release of OAuth2orize and not the likely outdated version in the main directory of this project.